+++ setup/tools/util.sh
@@ -2,7 +2,7 @@
 
 
 # * APT INSTALLED UTILS *
-apt-fast install -y nano tmux watch micro htop ftp inetutils-ping nfs-common
+apt-fast install --no-install-recommends -y nano tmux watch micro htop ftp inetutils-ping nfs-common
 
 # * GIT INSTALLED UTILS *